#refactor #javaDoc
