+++ fine-card-service/src/main/java/bank/service/finecardservice/client/common/CardType.java
@@ -1,0 +1,12 @@
+package bank.service.finecardservice.client.common;
+
+/**
+ * The enum Card type.
+ *
+ * @author Ruslan Potapov
+ */
+public enum CardType {
+    FINES,
+    REGULAR,
+    ADVISE
+}
