--- fine-card-service/src/main/java/bank/service/finecardservice/client/dto/CardType.java
@@ -1,7 +1,0 @@
-package bank.service.finecardservice.client.dto;
-
-public enum CardType {
-    FINES,
-    REGULAR,
-    ADVISE
-}
