+++ fine-card-service/src/main/java/bank/service/finecardservice/client/dto/FinesResponse.java
@@ -2,5 +2,10 @@ package bank.service.finecardservice.client.dto;
 
 import java.util.ArrayList;
 
+/**
+ * The type Fines response.
+ *
+ * @author Ruslan Potapov
+ */
 public class FinesResponse extends ArrayList<FineDTO> {
 }
