+++ fine-card-service/src/main/java/bank/service/finecardservice/configuration/FinesProperties.java
@@ -4,6 +4,10 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
+/**
+ * The type Fines properties.
+ * @author Ruslan Potapov
+ */
 @Data
 @NoArgsConstructor
 @ConfigurationProperties(prefix = "fines")
