+++ fine-card-service/src/main/java/bank/service/finecardservice/service/FinesService.java
@@ -2,9 +2,14 @@ package bank.service.finecardservice.service;
 
 
 import bank.service.finecardservice.entity.FineCard;
-import bank.service.finecardservice.client.dto.UserData;
+import bank.service.finecardservice.client.common.UserData;
 import reactor.core.publisher.Flux;
 
+/**
+ * The interface Fines service.
+ *
+ * @author Ruslan Potpaov
+ */
 public interface FinesService {
     Flux<FineCard> loadFines(UserData userData);
 }
