+++ fine-card-service/src/main/java/bank/service/finecardservice/service/RemoteFinesService.java
@@ -1,26 +1,30 @@
 package bank.service.finecardservice.service;
 
 
-import bank.service.finecardservice.client.GovermentClient;
+import bank.service.finecardservice.client.GovernmentClient;
+import bank.service.finecardservice.client.common.CardType;
+import bank.service.finecardservice.client.common.UserData;
 import bank.service.finecardservice.configuration.FinesProperties;
-import bank.service.finecardservice.client.dto.CardType;
 import bank.service.finecardservice.entity.FineCard;
-import bank.service.finecardservice.client.dto.UserData;
 import lombok.RequiredArgsConstructor;
-
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 
+/**
+ * The type Remote fines service.
+ *
+ * @author Ruslan Potapov
+ */
 @Service
 @RequiredArgsConstructor
 public class RemoteFinesService implements FinesService {
     private final FinesProperties properties;
-    private final GovermentClient GovermentClient;
+    private final GovernmentClient governmentClient;
 
     @Override
     public Flux<FineCard> loadFines(UserData userData) {
-        return GovermentClient.getFines(userData.getUserId())
-                .flatMapIterable(res-> res)
+        return governmentClient.getFines(userData.getUserId())
+                .flatMapIterable(res -> res)
                 .map(fine ->
                         FineCard.builder()
                                 .userId(userData.getUserId())
